+++ src/Installer/PackagesInstaller.php
@@ -12,6 +12,7 @@ namespace Youwe\TestingSuite\Composer\Installer;
 use Composer\Composer;
 use Composer\IO\IOInterface;
 use Youwe\Composer\DependencyInstaller\DependencyInstaller;
+use Youwe\TestingSuite\Composer\MappingResolver;
 use Youwe\TestingSuite\Composer\ProjectTypeResolver;
 
 /**
@@ -33,23 +34,23 @@ class PackagesInstaller implements InstallerInterface
 
     /** @var array */
     private $mapping = [
-        'default' => [],
+        MappingResolver::DEFAULT_MAPPING_TYPE => [],
         'magento1' => [
             [
                 'name' => 'youwe/coding-standard-magento1',
-                'version' => '@stable'
+                'version' => '^1.3.0'
             ]
         ],
         'magento2' => [
             [
                 'name' => 'youwe/coding-standard-magento2',
-                'version' => '@stable'
+                'version' => '^1.7.0'
             ]
         ],
         'laravel' => [
             [
                 'name' => 'elgentos/laravel-coding-standard',
-                'version' => '@stable'
+                'version' => '^1.0.0'
             ]
         ]
     ];