+++ src/MappingResolver.php
@@ -17,6 +17,8 @@ class MappingResolver
     /** @var ProjectTypeResolver */
     private $typeResolver;
 
+    public const DEFAULT_MAPPING_TYPE = 'default';
+
     /**
      * Constructor.
      *