+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Old pathing to Mediact pathing in files `phpcs.xml`, `phpmd.xml` & `grumphp.yml` will now be replaced by 
   Youwe pathing to prevent error of phpcs/phpmd/grumphp.
 
+### Changed
+- Versions of packages required are now not using `@stable` anymore.
+
 ## 2.10.0 - 2021-03-10
 ### Added
 - Copyright.